More sub-epoch fixes

+++ src/types/sub_epoch_summary.py
@@ -12,5 +12,5 @@ class SubEpochSummary(Streamable):
     prev_subepoch_summary_hash: bytes32
     reward_chain_hash: bytes32  # hash of reward chain at end of last segment
     num_sub_blocks_overflow: uint8  # How many more sub-blocks than 384*(N-1)
-    new_difficulty: Optional[uint128]  # Only once per epoch (diff adjustment)
+    new_difficulty: Optional[uint64]  # Only once per epoch (diff adjustment)
     new_ips: Optional[uint64]  # Only once per epoch (diff adjustment)
